requirements.txt
elevenlabs==2.9.2
openai==1.99.2
requests==2.32.4
moviepy==1.0.3
Pillow>=9.0.0
numpy>=1.21.0
flask==3.0.0
flask-cors==4.0.0
